Implement drawing.

+++ src/chip8/video.hpp
@@ -7,4 +7,4 @@
  *  Chip-8 displays a 64x32 screen.
  */
 
-using VideoBuf = std::array<uint8_t, 64*32>;
+using VideoBuf = std::array<uint32_t, 64*32>;
